don't use fetch on the server
In some cases, localhost:3000 won't be where the api is deployed to and
excluding the origin is not allowed when its called from the server.
More specifically, doing this restructure allows the prisma-examples to
be deployed more easily on vercel.  In Vercel, The nextjs server is
deployed to run on edge functions, the origin of which is not known at
deploy time.